Include both outpack and orderly packages when building wheels.
When installing this repo with pip, it would only install the outpack
package, not the orderly one. It seems that by default, hatch looks for
`src/$PROJECT_NAME` to figure out what to build, which in our case is
just the outpack package.

Thankfully it offers an override that allows multiple package paths to
be listed, which we can use to include the two packages.